requirements.txt
django
python-dotenv
langchain
langchain-community
langchain-google-genai
pdfplumber
docx2txt
faiss-cpu
sentence-transformers
pypdf
python-magic
neo4j
langchain-neo4j
langchain-huggingface